requirement.txt
attrs==21.4.0
-e git+ssh://git@github.com/Olamshin/busybee.git@2045061ffdf8a05cf3b96b486112ffaaaeb35f76#egg=busybee
certifi==2021.10.8
charset-normalizer==2.0.12
cmd2==2.4.0
idna==3.3
pyperclip==1.8.2
PyYAML==6.0
requests==2.27.1
urllib3==1.26.8
wcwidth==0.2.5
asyncio==3.4.3
blinker==1.6.3
